Refactor conductivity sensor UI
